+++ Changelog.md
@@ -105,6 +105,8 @@
   You can now control the height of the `GalleryPanel` inside `InRibbonGallery` by setting `GalleryPanelContainerHeight`.
   - [#578](../../issues/578) - Theming of selected context ribbon tab  
   You can now use `Fluent.Ribbon.Brushes.RibbonWindow.TitleBackground`, `RibbonContextualTabGroup.TabItemSelectedForeground` and `RibbonContextualTabGroup.TabItemMouseOverForeground` to further control colors.
+  - [#590](../../issues/590) - SplitButton custom KeyTip  
+  You can now use `PrimaryActionKeyTipPostfix` and `SecondaryActionKeyTipPostfix` on `SplitButton` to control the postfix for key tips.
 
 ## 6.1.0